partial fix to visual artifact
the fact of resizing quickly causing a mismatch with topline is still present. A workaround with debouncing plus discrete delta logic has been implmenet, regardless the issue remains. Chances are it has something to do with ioctl, so not exactly sure how to work around that besides adding even more clamps?